+++ maps/services/FramesService.py
@@ -1,0 +1,57 @@
+import cv2
+import numpy as np
+import os
+from datetime import timedelta
+
+SAVING_FRAMES_PER_SECOND = 1
+
+
+def format_timedelta(td):
+    result = str(td)
+    try:
+        result, ms = result.split(".")
+    except ValueError:
+        return result + ".00".replace(":", "-")
+    ms = int(ms)
+    ms = round(ms / 1e4)
+    return f"{result}.{ms:02}".replace(":", "-")
+
+
+def get_saving_frames_durations(cap, saving_fps):
+    s = []
+    print(cap.get(cv2.CAP_PROP_FRAME_COUNT))
+    print(cap.get(cv2.CAP_PROP_FPS))
+    clip_duration = cap.get(cv2.CAP_PROP_FRAME_COUNT) / cap.get(cv2.CAP_PROP_FPS)
+    for i in np.arange(0, clip_duration, 1 / saving_fps):
+        s.append(i)
+    return s
+
+
+def create_frames():
+    video_file = 'videos/data-record.mp4'
+    filename = 'video-images'
+    filename += "-opencv"
+    if not os.path.isdir(filename):
+        os.mkdir(filename)
+    cap = cv2.VideoCapture(video_file)
+    fps = cap.get(cv2.CAP_PROP_FPS)
+    saving_frames_per_second = min(fps, SAVING_FRAMES_PER_SECOND)
+    saving_frames_durations = get_saving_frames_durations(cap, saving_frames_per_second)
+    count = 0
+    while True:
+        is_read, frame = cap.read()
+        if not is_read:
+            break
+        frame_duration = count / fps
+        try:
+            closest_duration = saving_frames_durations[0]
+        except IndexError:
+            break
+        if frame_duration >= closest_duration:
+            frame_duration_formatted = format_timedelta(timedelta(seconds=frame_duration))
+            cv2.imwrite(os.path.join(filename, f"frame{frame_duration_formatted}.jpg"), frame)
+            try:
+                saving_frames_durations.pop(0)
+            except IndexError:
+                pass
+        count += 1
