+++ maps/urls.py
@@ -3,8 +3,8 @@ from django.urls import path
 from . import views
 
 urlpatterns = [
-    path('', views.index, name='index'),
-    path('create_video', views.create_video, name='test'),
-    path('counting', views.counting, name='counting'),
+    path('', views.index_view, name='index'),
+    path('create_video', views.create_video_view, name='test'),
+    path('counting', views.counting_view, name='counting'),
     path('get_osm_data', views.get_osm_data, name='get_osm_data'),
 ]
